Make: Match directory structure of SimSIMD
