+++ package.json
@@ -20,7 +20,7 @@
     "node-addon-api": "^3.0.0"
   },
   "scripts": {
-    "test": "node --test ./javascript/test"
+    "test": "node --test ./scripts/test.js"
   },
   "devDependencies": {
     "@semantic-release/exec": "^6.0.3",
